refactor(i18n): change local folder name
