+++ .storybook/preview.tsx
@@ -4,7 +4,7 @@ import { createGlobalStyle, ThemeProvider } from 'styled-components';
 import 'typeface-lato';
 import 'typeface-sedgwick-ave';
 import theme from 'app/theme';
-import 'locales';
+import 'i18n';
 
 const Global = createGlobalStyle`
   body {
