+++ src/app/profiles/index.tsx
@@ -2,6 +2,6 @@ import './browser';
 import React from 'react';
 import { render } from 'react-dom';
 import App from './App';
-import 'locales';
+import 'i18n';
 
 render(<App />, document.getElementById('root'));
