+++ src/i18n/resources/fr/index.ts
