+++ src/i18n/resources/index.ts
